updated selectedPortfolio reference to indicate possible absence, isDefault will default to true if no default portfolio has been set up, further unified method naming scheme, simplified optional completion calls, resolved cascading by guard statements, added no portfolio indicator, transaction detail view continuously updates to reflect current price, created custom transacactionCell fo future replacement, added pull to refresh on transaction table controller

+++ Krypton/PortfolioDetailController.swift
@@ -146,6 +146,10 @@ class PortfolioDetailController: UITableViewController {
     }
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        guard indexPath != aliasIndexPath, indexPath != isDefaultIndexPath else {
+            return
+        }
+        
         if portfolio.storedAddresses.count > 0 && indexPath.section == 2 || portfolio.storedAddresses.count == 0 && indexPath.section == 1 {
             deletePortfolio()
         } else {
